Combine sync and manager into PayjoinManager
