+++ lib/locator.dart
@@ -9,8 +9,8 @@ import 'package:bb_mobile/_pkg/launcher.dart';
 import 'package:bb_mobile/_pkg/logger.dart';
 import 'package:bb_mobile/_pkg/mempool_api.dart';
 import 'package:bb_mobile/_pkg/nfc.dart';
+import 'package:bb_mobile/_pkg/payjoin/manager.dart';
 import 'package:bb_mobile/_pkg/payjoin/session_storage.dart';
-import 'package:bb_mobile/_pkg/payjoin/sync.dart';
 import 'package:bb_mobile/_pkg/storage/hive.dart';
 import 'package:bb_mobile/_pkg/storage/secure_storage.dart';
 import 'package:bb_mobile/_pkg/storage/storage.dart';
@@ -243,7 +243,12 @@ Future _setupWalletServices() async {
     ),
   );
 
-  locator.registerFactory<PayjoinSync>(() => PayjoinSync());
+  locator.registerSingleton<PayjoinManager>(
+    PayjoinManager(
+      locator<PayjoinSessionStorage>(),
+      locator<WalletTx>(),
+    ),
+  );
 }
 
 Future _setupBlocs() async {
